refactor: update ME for TDP template to use adaptiveCard (ts/js) (#10705)
* refactor: use adaptive card

* refactor: revert

* refactor: more clean

* refactor: ts

* refactor: js rename
