+++ templates/js/message-extension/src/adaptiveCards/linkUnfurlingCard.json
@@ -1,0 +1,45 @@
+{
+    "type": "AdaptiveCard",
+    "body": [
+      {
+        "type": "ColumnSet",
+        "columns": [
+          {
+            "type": "Column",
+            "width": "1",
+            "items": [
+              {
+                "type": "Image",
+                "url": "https://raw.githubusercontent.com/microsoft/botframework-sdk/master/icon.png"
+              }
+            ]
+          },
+          {
+            "type": "Column",
+            "width": "2",
+            "items": [
+              {
+                "type": "TextBlock",
+                "size": "Large",
+                "weight": "Bolder",
+                "text": "Link Unfurling card"
+              },
+              {
+                "type": "TextBlock",
+                "text": "This a sample card for Link Unfurling"
+              }
+            ]
+          }
+        ]
+      }
+    ],
+    "actions": [
+      {
+        "type": "Action.OpenUrl",
+        "title": "Link Unfurling Docs",
+        "url": "https://learn.microsoft.com/en-us/microsoftteams/platform/messaging-extensions/how-to/link-unfurling"
+      }
+    ],
+    "$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
+    "version": "1.4"
+  }
