+++ templates/js/message-extension/src/config.js
