+++ templates/js/message-extension/src/index.js
