+++ templates/js/message-extension/src/teamsBot.js
@@ -1,0 +1,91 @@
+const axios = require("axios");
+const querystring = require("querystring");
+const { TeamsActivityHandler, CardFactory } = require("botbuilder");
+const ACData = require("adaptivecards-templating");
+const searchResultCard = require("./adaptiveCards/searchResultCard.json");
+const actionCard = require("./adaptiveCards/actionCard.json");
+const linkUnfurlingCard = require("./adaptiveCards/linkUnfurlingCard.json");
+
+class TeamsBot extends TeamsActivityHandler {
+  // Action.
+  handleTeamsMessagingExtensionSubmitAction(context, action) {
+    // The user has chosen to create a card by choosing the 'Create Card' context menu command.
+    const template = new ACData.Template(actionCard);
+    const card = template.expand({
+      $root: {
+        title: action.data.title ?? "",
+        subTitle: action.data.subTitle ?? "",
+        text: action.data.text ?? "",
+      },
+    });
+    const attachment = CardFactory.adaptiveCard(card);
+    return {
+      composeExtension: {
+        type: "result",
+        attachmentLayout: "list",
+        attachments: [attachment],
+      },
+    };
+  }
+
+  // Search.
+  async handleTeamsMessagingExtensionQuery(context, query) {
+    const searchQuery = query.parameters[0].value;
+    const response = await axios.get(
+      `http://registry.npmjs.com/-/v1/search?${querystring.stringify({
+        text: searchQuery,
+        size: 8,
+      })}`
+    );
+
+    const attachments = [];
+    response.data.objects.forEach((obj) => {
+      const template = new ACData.Template(searchResultCard);
+      const card = template.expand({
+        $root: {
+          name: obj.package.name,
+          description: obj.package.description,
+        },
+      });
+      const preview = CardFactory.heroCard(obj.package.name);
+      const attachment = { ...CardFactory.adaptiveCard(card), preview };
+      attachments.push(attachment);
+    });
+
+    return {
+      composeExtension: {
+        type: "result",
+        attachmentLayout: "list",
+        attachments: attachments,
+      },
+    };
+  }
+
+  // Link Unfurling.
+  handleTeamsAppBasedLinkQuery(context, query) {
+    const previewCard = CardFactory.thumbnailCard("Preview Card", query.url, [
+      "https://raw.githubusercontent.com/microsoft/botframework-sdk/master/icon.png",
+    ]);
+
+    const attachment = { ...CardFactory.adaptiveCard(linkUnfurlingCard), preview: previewCard };
+
+    return {
+      composeExtension: {
+        type: "result",
+        attachmentLayout: "list",
+        attachments: [attachment],
+        suggestedActions: {
+          actions: [
+            {
+              title: "default",
+              type: "setCachePolicy",
+              value: '{"type":"no-cache"}',
+            },
+          ],
+        },
+      },
+    };
+  }
+}
+
+module.exports.TeamsBot = TeamsBot;
