+++ templates/ts/message-extension/src/adaptiveCards/searchResultCard.json
@@ -1,0 +1,19 @@
+{
+    "type": "AdaptiveCard",
+    "body": [
+      {
+        "type": "TextBlock",
+        "text": "${name}",
+        "wrap": true,
+        "size": "Large"
+      },
+      {
+        "type": "TextBlock",
+        "text": "${description}",
+        "wrap": true,
+        "size": "Medium"
+      }
+    ],
+    "$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
+    "version": "1.4"
+}
