+++ templates/ts/message-extension/src/config.ts
