+++ templates/ts/message-extension/src/index.ts
