+++ templates/ts/message-extension/src/teamsBot.ts
@@ -1,0 +1,108 @@
+import { default as axios } from "axios";
+import * as querystring from "querystring";
+import {
+  TeamsActivityHandler,
+  CardFactory,
+  TurnContext,
+  MessagingExtensionAction,
+  MessagingExtensionQuery,
+  MessagingExtensionResponse,
+  MessagingExtensionActionResponse,
+  AppBasedLinkQuery,
+} from "botbuilder";
+import * as ACData from "adaptivecards-templating";
+import searchResultCard from "./adaptiveCards/searchResultCard.json";
+import linkUnfurlingCard from "./adaptiveCards/linkUnfurlingCard.json";
+import actionCard from "./adaptiveCards/actionCard.json";
+
+export class TeamsBot extends TeamsActivityHandler {
+  // Action.
+  public async handleTeamsMessagingExtensionSubmitAction(
+    context: TurnContext,
+    action: MessagingExtensionAction
+  ): Promise<MessagingExtensionActionResponse> {
+    // The user has chosen to create a card by choosing the 'Create Card' context menu command.
+    const template = new ACData.Template(actionCard);
+    const card = template.expand({
+      $root: {
+        title: action.data.title ?? "",
+        subTitle: action.data.subTitle ?? "",
+        text: action.data.text ?? "",
+      },
+    });
+    const attachment = CardFactory.adaptiveCard(card);
+    return {
+      composeExtension: {
+        type: "result",
+        attachmentLayout: "list",
+        attachments: [attachment],
+      },
+    };
+  }
+
+  // Search.
+  public async handleTeamsMessagingExtensionQuery(
+    context: TurnContext,
+    query: MessagingExtensionQuery
+  ): Promise<MessagingExtensionResponse> {
+    const searchQuery = query.parameters[0].value;
+    const response = await axios.get(
+      `http://registry.npmjs.com/-/v1/search?${querystring.stringify({
+        text: searchQuery,
+        size: 8,
+      })}`
+    );
+
+    const attachments = [];
+    response.data.objects.forEach((obj) => {
+      const template = new ACData.Template(searchResultCard);
+      const card = template.expand({
+        $root: {
+          name: obj.package.name,
+          description: obj.package.description,
+        },
+      });
+      const preview = CardFactory.heroCard(obj.package.name);
+      const attachment = { ...CardFactory.adaptiveCard(card), preview };
+      attachments.push(attachment);
+    });
+
+    return {
+      composeExtension: {
+        type: "result",
+        attachmentLayout: "list",
+        attachments: attachments,
+      },
+    };
+  }
+
+  // Link Unfurling.
+  public async handleTeamsAppBasedLinkQuery(
+    context: TurnContext,
+    query: AppBasedLinkQuery
+  ): Promise<MessagingExtensionResponse> {
+    // When the returned card is an adaptive card, the previewCard property of the attachment is required.
+    const previewCard = CardFactory.thumbnailCard("Preview Card", query.url, [
+      "https://raw.githubusercontent.com/microsoft/botframework-sdk/master/icon.png",
+    ]);
+
+    const attachment = { ...CardFactory.adaptiveCard(linkUnfurlingCard), preview: previewCard };
+
+    return {
+      composeExtension: {
+        type: "result",
+        attachmentLayout: "list",
+        attachments: [attachment],
+        suggestedActions: {
+          actions: [
+            {
+              title: "default",
+              type: "setCachePolicy",
+              value: '{"type":"no-cache"}',
+            },
+          ],
+        },
+      },
+    };
+  }
+}
