--- templates/ts/message-extension/teamsBot.ts
@@ -1,181 +1,0 @@
-import { default as axios } from "axios";
-import * as querystring from "querystring";
-import {
-  TeamsActivityHandler,
-  CardFactory,
-  TurnContext,
-  MessagingExtensionAction,
-  MessagingExtensionQuery,
-  MessagingExtensionResponse,
-  MessagingExtensionActionResponse,
-  AppBasedLinkQuery,
-} from "botbuilder";
-
-export class TeamsBot extends TeamsActivityHandler {
-  // Action.
-  public async handleTeamsMessagingExtensionSubmitAction(
-    context: TurnContext,
-    action: MessagingExtensionAction
-  ): Promise<MessagingExtensionActionResponse> {
-    switch (action.commandId) {
-      case "createCard":
-        return createCardCommand(context, action);
-      case "shareMessage":
-        return shareMessageCommand(context, action);
-      default:
-        throw new Error("NotImplemented");
-    }
-  }
-
-  // Search.
-  public async handleTeamsMessagingExtensionQuery(
-    context: TurnContext,
-    query: MessagingExtensionQuery
-  ): Promise<MessagingExtensionResponse> {
-    const searchQuery = query.parameters[0].value;
-    const response = await axios.get(
-      `http://registry.npmjs.com/-/v1/search?${querystring.stringify({
-        text: searchQuery,
-        size: 8,
-      })}`
-    );
-
-    const attachments = [];
-    response.data.objects.forEach((obj) => {
-      const heroCard = CardFactory.heroCard(obj.package.name);
-      const preview = CardFactory.heroCard(obj.package.name);
-      preview.content.tap = {
-        type: "invoke",
-        value: { name: obj.package.name, description: obj.package.description },
-      };
-      const attachment = { ...heroCard, preview };
-      attachments.push(attachment);
-    });
-
-    return {
-      composeExtension: {
-        type: "result",
-        attachmentLayout: "list",
-        attachments: attachments,
-      },
-    };
-  }
-
-  public async handleTeamsMessagingExtensionSelectItem(
-    context: TurnContext,
-    obj: any
-  ): Promise<MessagingExtensionResponse> {
-    return {
-      composeExtension: {
-        type: "result",
-        attachmentLayout: "list",
-        attachments: [CardFactory.heroCard(obj.name, obj.description)],
-      },
-    };
-  }
-
-  // Link Unfurling.
-  public async handleTeamsAppBasedLinkQuery(
-    context: TurnContext,
-    query: AppBasedLinkQuery
-  ): Promise<MessagingExtensionResponse> {
-    const attachment = CardFactory.thumbnailCard("Image Preview Card", query.url, [query.url]);
-
-    // By default the link unfurling result is cached in Teams for 30 minutes.
-    // The code has set a cache policy and removed the cache for the app. Learn more here: https://learn.microsoft.com/microsoftteams/platform/messaging-extensions/how-to/link-unfurling?tabs=dotnet%2Cadvantages#remove-link-unfurling-cache
-    return {
-      composeExtension: {
-        type: "result",
-        attachmentLayout: "list",
-        attachments: [attachment],
-        suggestedActions: {
-          actions: [
-            {
-              title: "default",
-              type: "setCachePolicy",
-              value: '{"type":"no-cache"}',
-            },
-          ],
-        },
-      },
-    };
-  }
-}
-
-async function createCardCommand(
-  context: TurnContext,
-  action: MessagingExtensionAction
-): Promise<MessagingExtensionResponse> {
-  // The user has chosen to create a card by choosing the 'Create Card' context menu command.
-  const data = action.data;
-  const heroCard = CardFactory.heroCard(data.title, data.text);
-  heroCard.content.subtitle = data.subTitle;
-  const attachment = {
-    contentType: heroCard.contentType,
-    content: heroCard.content,
-    preview: heroCard,
-  };
-
-  return {
-    composeExtension: {
-      type: "result",
-      attachmentLayout: "list",
-      attachments: [attachment],
-    },
-  };
-}
-
-async function shareMessageCommand(
-  context: TurnContext,
-  action: MessagingExtensionAction
-): Promise<MessagingExtensionResponse> {
-  // The user has chosen to share a message by choosing the 'Share Message' context menu command.
-  let userName = "unknown";
-  if (
-    action.messagePayload &&
-    action.messagePayload.from &&
-    action.messagePayload.from.user &&
-    action.messagePayload.from.user.displayName
-  ) {
-    userName = action.messagePayload.from.user.displayName;
-  }
-
-  // This Message Extension example allows the user to check a box to include an image with the
-  // shared message.  This demonstrates sending custom parameters along with the message payload.
-  let images = [];
-  const includeImage = action.data.includeImage;
-  if (includeImage === "true") {
-    images = [
-      "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQtB3AwMUeNoq4gUBGe6Ocj8kyh3bXa9ZbV7u1fVKQoyKFHdkqU",
-    ];
-  }
-  const heroCard = CardFactory.heroCard(
-    `${userName} originally sent this message:`,
-    action.messagePayload.body.content,
-    images
-  );
-
-  if (
-    action.messagePayload &&
-    action.messagePayload.attachments &&
-    action.messagePayload.attachments.length > 0
-  ) {
-    // This sample does not add the MessagePayload Attachments.  This is left as an
-    // exercise for the user.
-    heroCard.content.subtitle = `(${action.messagePayload.attachments.length} Attachments not included)`;
-  }
-
-  const attachment = {
-    contentType: heroCard.contentType,
-    content: heroCard.content,
-    preview: heroCard,
-  };
-
-  return {
-    composeExtension: {
-      type: "result",
-      attachmentLayout: "list",
-      attachments: [attachment],
-    },
-  };
-}
